requirements.txt
pyyaml
zeroc-ice
